Version 1.2.0.0
Fixed a few issues and tweaked app design.

Fixed an issue where the invoker of the app would have their desktop checked instead of the currently logged in user.

Fixed an issue where ShortcutFile.Delete method would not throw errors back to the main window to be handled.

Changed ShortcutFile.Delete method to return type void instead of bool.

Changed the layout of the app slightly to allow for more room for shortcut list to show items.

Renamed action buttons' text to say "shortcut" instead of "item".

Changed close button to use bitmap icon instead of text.

Added to list of shortcuts a way to identify which desktop the item is located on.