Version 1.3.0.0
Added CurrentUserInfo class that manages detecting the current user and their desktop directory via Windows shell calls.

DesktopShortcutManager class no longer attempts to locate the Desktop directory. Instead this is directory is passed as a string argument in the contructor.

Fixed a spelling/grammer mistake of a member variable in DesktopShortcutManger.

Warn user when invoking the app as not the SYSTEM user on app start. Allow them to decide if they want to proceed.

If invoked as not the SYSTEM user, display banner and have option to show pop message explaining why things may not work.